+++ examples/features/src/big_compute_buffers/mod.rs
@@ -85,7 +85,7 @@ pub async fn execute_gpu_inner(
     let mut data = Vec::new();
     for staging_buffer in &staging_buffers {
         let slice = staging_buffer.slice(..);
-        let mapped = slice.get_mapped_range();
+        let mapped = slice.get_mapped_range().unwrap();
         let chunk: Vec<f32> = bytemuck::allocation::pod_collect_to_vec(&mapped);
         data.extend_from_slice(&chunk);
         drop(mapped);