+++ examples/features/src/repeated_compute/mod.rs
@@ -115,7 +115,7 @@ async fn compute(local_buffer: &mut [u32], context: &WgpuContext) {
     log::info!("Result received.");
     // NOW we can call get_mapped_range.
     {
-        let view = buffer_slice.get_mapped_range();
+        let view = buffer_slice.get_mapped_range().unwrap();
         let data: Vec<u32> = bytemuck::allocation::pod_collect_to_vec(&view);
         local_buffer.copy_from_slice(&data);
     }