+++ tests/tests/wgpu-gpu/shader/array_size_overrides.rs
@@ -156,7 +156,7 @@ async fn array_size_overrides(
     mapping_buffer.slice(..).map_async(MapMode::Read, |_| ());
     ctx.async_poll(PollType::wait_indefinitely()).await.unwrap();
 
-    let mapped = mapping_buffer.slice(..).get_mapped_range();
+    let mapped = mapping_buffer.slice(..).get_mapped_range().unwrap();
 
     let typed: &[u32] = bytemuck::cast_slice(&mapped);
     assert_eq!(typed, out);