+++ tests/tests/wgpu-gpu/shader/zero_init_workgroup_mem.rs
@@ -144,7 +144,7 @@ static ZERO_INIT_WORKGROUP_MEMORY: GpuTestConfiguration = GpuTestConfiguration::
         mapping_buffer.slice(..).map_async(MapMode::Read, |_| ());
         ctx.async_poll(PollType::wait_indefinitely()).await.unwrap();
 
-        let mapped = mapping_buffer.slice(..).get_mapped_range();
+        let mapped = mapping_buffer.slice(..).get_mapped_range().unwrap();
 
         let typed: &[u32] = bytemuck::cast_slice(&mapped);